+++ ts/src/game/component/job/job.ts
@@ -1,8 +1,8 @@
-import { InvalidStateError } from "../../common/error/invalidStateError.js";
-import { NotInitializedError } from "../../common/error/notInitializedError.js";
-import { RenderContext } from "../../rendering/renderContext.js";
-import { JobOwner } from "../component/job/jobOwner.js";
-import { Entity } from "../entity/entity.js";
+import { InvalidStateError } from "../../../common/error/invalidStateError.js";
+import { NotInitializedError } from "../../../common/error/notInitializedError.js";
+import { RenderContext } from "../../../rendering/renderContext.js";
+import { Entity } from "../../entity/entity.js";
+import { JobOwner } from "./jobOwner.js";
 import { JobConstraint } from "./jobConstraint.js";
 
 export enum JobState {
