+++ ts/src/game/component/job/jobAddedEvent.ts
@@ -1,5 +1,5 @@
-import { Job } from "../../job/job.js";
 import { ComponentEvent } from "../componentEvent.js";
+import { Job } from "./job.js";
 import { JobQueueComponent } from "./jobQueueComponent.js";
 
 export class JobAddedEvent extends ComponentEvent<JobQueueComponent> {
