+++ ts/src/game/component/job/jobConstraint.ts
@@ -1,4 +1,4 @@
-import { Entity } from "../entity/entity.js";
+import { Entity } from "../../entity/entity.js";
 import { Job } from "./job.js";
 
 export interface JobConstraint {
