+++ ts/src/game/component/job/jobConstraintsError.ts
