+++ ts/src/game/component/job/jobOwner.ts
@@ -1,4 +1,4 @@
-import { Job } from "../../job/job.js";
+import { Job } from "./job.js";
 
 export interface JobOwner {
     /**
