+++ ts/src/game/component/job/jobRunnerComponent.ts
@@ -1,7 +1,7 @@
 import { Point } from "../../../common/point.js";
 import { RenderContext } from "../../../rendering/renderContext.js";
-import { Job } from "../../job/job.js";
 import { EntityComponent } from "../entityComponent.js";
+import { Job } from "./job.js";
 import { JobOwner } from "./jobOwner.js";
 import { JobQueueComponent } from "./jobQueueComponent.js";
