+++ ts/src/game/component/job/jobs/attack/attackJob.ts
