+++ ts/src/game/component/job/jobs/chopTreeJob.ts
@@ -1,18 +1,18 @@
-import { sprites2 } from "../../../asset/sprite.js";
-import { generateId } from "../../../common/idGenerator.js";
-import { woodResourceItem } from "../../../data/inventory/resources.js";
-import { RenderContext } from "../../../rendering/renderContext.js";
-import { BlinkingImageAnimation } from "../../../rendering/visual/blinkingImageAnimation.js";
-import { HealthComponent } from "../../component/health/healthComponent.js";
-import { TreeComponent } from "../../component/resource/treeComponent.js";
-import { InventoryComponent } from "../../component/inventory/inventoryComponent.js";
-import { TilesComponent } from "../../component/tile/tilesComponent.js";
-import { Entity } from "../../entity/entity.js";
-import { treePrefab } from "../../prefab/treePrefab.js";
-import { SelectedEntityItem } from "../../selection/selectedEntityItem.js";
-import { SelectedTileItem } from "../../selection/selectedTileItem.js";
-import { SelectedWorldItem } from "../../selection/selectedWorldItem.js";
-import { TileSize } from "../../tile/tile.js";
+import { sprites2 } from "../../../../asset/sprite.js";
+import { generateId } from "../../../../common/idGenerator.js";
+import { woodResourceItem } from "../../../../data/inventory/resources.js";
+import { RenderContext } from "../../../../rendering/renderContext.js";
+import { BlinkingImageAnimation } from "../../../../rendering/visual/blinkingImageAnimation.js";
+import { HealthComponent } from "../../../component/health/healthComponent.js";
+import { TreeComponent } from "../../../component/resource/treeComponent.js";
+import { InventoryComponent } from "../../../component/inventory/inventoryComponent.js";
+import { TilesComponent } from "../../../component/tile/tilesComponent.js";
+import { Entity } from "../../../entity/entity.js";
+import { treePrefab } from "../../../prefab/treePrefab.js";
+import { SelectedEntityItem } from "../../../selection/selectedEntityItem.js";
+import { SelectedTileItem } from "../../../selection/selectedTileItem.js";
+import { SelectedWorldItem } from "../../../selection/selectedWorldItem.js";
+import { TileSize } from "../../../tile/tile.js";
 import { Job } from "../job.js";
 import { JobConstraintsError } from "../jobConstraintsError.js";
 import { MoveToBeforeJob } from "./moveToBeforeJob.js";
