+++ ts/src/game/component/job/jobs/moveToBeforeJob.ts
@@ -1,11 +1,11 @@
-import { InvalidStateError } from "../../../common/error/invalidStateError.js";
+import { InvalidStateError } from "../../../../common/error/invalidStateError.js";
 import {
     isPointAdjacentTo,
     Point,
     pointEquals,
-} from "../../../common/point.js";
-import { PathFindingComponent } from "../../component/root/path/pathFindingComponent.js";
-import { PathResultStatus } from "../../component/root/path/pathResult.js";
+} from "../../../../common/point.js";
+import { PathFindingComponent } from "../../../component/root/path/pathFindingComponent.js";
+import { PathResultStatus } from "../../../component/root/path/pathResult.js";
 import { Job } from "../job.js";
 import { JobConstraint } from "../jobConstraint.js";
 import { JobConstraintsError } from "../jobConstraintsError.js";
