+++ ts/src/game/component/job/multipleStepJob.ts
@@ -1,6 +1,6 @@
-import { InvalidStateError } from "../../common/error/invalidStateError.js";
-import { RenderContext } from "../../rendering/renderContext.js";
-import { JobOwner } from "../component/job/jobOwner.js";
+import { InvalidStateError } from "../../../common/error/invalidStateError.js";
+import { RenderContext } from "../../../rendering/renderContext.js";
+import { JobOwner } from "./jobOwner.js";
 import { Job, JobState } from "./job.js";
 
 /**
