+++ ts/src/game/component/job/query/selectedItemIsTargetQuery.ts
@@ -1,8 +1,8 @@
-import { Job } from "../../../job/job.js";
-import { ChopTreeJob } from "../../../job/jobs/chopTreeJob.js";
 import { SelectedEntityItem } from "../../../selection/selectedEntityItem.js";
 import { SelectedTileItem } from "../../../selection/selectedTileItem.js";
 import { SelectedWorldItem } from "../../../selection/selectedWorldItem.js";
+import { Job } from "../job.js";
+import { ChopTreeJob } from "../jobs/chopTreeJob.js";
 import { JobQuery } from "./jobQuery.js";
 
 /**
