+++ ts/src/game/interaction/state/building2/buildConfirmState.ts
@@ -8,9 +8,9 @@ import { uiBox } from "../../../../ui/dsl/uiBoxDsl.js";
 import { fillUiSize } from "../../../../ui/uiSize.js";
 import { InventoryComponent } from "../../../component/inventory/inventoryComponent.js";
 import { JobQueueComponent } from "../../../component/job/jobQueueComponent.js";
+import { BuildJob } from "../../../component/job/jobs/buildJob.js";
 import { ChunkMapComponent } from "../../../component/root/chunk/chunkMapComponent.js";
 import { TilesComponent } from "../../../component/tile/tilesComponent.js";
-import { BuildJob } from "../../../job/jobs/buildJob.js";
 import { buildingFactory } from "../../../prefab/buildingFactory.js";
 import { GroundTile } from "../../../tile/ground.js";
 import { TileSize } from "../../../tile/tile.js";
