+++ ts/src/game/interaction/state/resource/collectChestState.ts
@@ -7,8 +7,8 @@ import { fillUiSize, wrapUiSize } from "../../../../ui/uiSize.js";
 import { UIView } from "../../../../ui/uiView.js";
 import { UIFlowGrid } from "../../../../ui/view/uiFlowGrid.js";
 import { JobQueueComponent } from "../../../component/job/jobQueueComponent.js";
+import { CollectChestJob } from "../../../component/job/jobs/chest/collectChestJob.js";
 import { ChestComponent } from "../../../component/resource/chestComponent.js";
-import { CollectChestJob } from "../../../job/jobs/chest/collectChestJob.js";
 import { InteractionState } from "../../handler/interactionState.js";
 import { UIActionbarScaffold } from "../../view/actionbar/uiActionbarScaffold.js";
 import { UIBorderTitle } from "../../view/uiBorderTitle.js";
