+++ ts/src/game/interaction/state/selection/selectionState.ts
@@ -4,6 +4,7 @@ import { RenderContext } from "../../../../rendering/renderContext.js";
 import { uiBox } from "../../../../ui/dsl/uiBoxDsl.js";
 import { fillUiSize } from "../../../../ui/uiSize.js";
 import { WorkerBehaviorComponent } from "../../../component/behavior/workerBehaviorComponent.js";
+import { Job } from "../../../component/job/job.js";
 import { JobQueueComponent } from "../../../component/job/jobQueueComponent.js";
 import { query } from "../../../component/job/query/jobQuery.js";
 import { SelectedItemIsTargetQuery } from "../../../component/job/query/selectedItemIsTargetQuery.js";
@@ -11,7 +12,6 @@ import { ChestComponent } from "../../../component/resource/chestComponent.js";
 import { TreeComponent } from "../../../component/resource/treeComponent.js";
 import { ChunkMapComponent } from "../../../component/root/chunk/chunkMapComponent.js";
 import { TilesComponent } from "../../../component/tile/tilesComponent.js";
-import { Job } from "../../../job/job.js";
 import { SelectedEntityItem } from "../../../selection/selectedEntityItem.js";
 import { SelectedTileItem } from "../../../selection/selectedTileItem.js";
 import { SelectedWorldItem } from "../../../selection/selectedWorldItem.js";
