+++ ts/test/game/component/job/jobRunnerComponentTest.ts
@@ -1,8 +1,8 @@
 import * as assert from "node:assert";
 import { describe, it } from "node:test";
 import { JobRunnerComponent } from "../../../../src/game/component/job/jobRunnerComponent.js";
-import { MoveJob } from "../../../../src/game/job/jobs/moveJob.js";
 import { Entity } from "../../../../src/game/entity/entity.js";
+import { MoveJob } from "../../../../src/game/component/job/jobs/moveJob.js";
 
 describe("JobRunnerComponent Tests", () => {
     it("Set active job to the assigned job", () => {
